deriv_portfolio_bot_dropin/requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.8.2
websockets==12.0
tenacity==9.0.0
numpy==2.0.2
pandas==2.2.2
matplotlib==3.9.0
typer==0.12.3
rich==13.7.1

# testing & quality
pytest==8.2.0
pytest-asyncio==0.24.0
mypy==1.11.1
black==24.8.0
isort==5.13.2
flake8==7.1.1
